Remove check for already explained transactions.
Our Chrome Extension is only active on URLs for *unexplained* transactions and
so this check is unnecessary.

Although this change makes the JS less useful when used as a bookmarklet, we're
about to introduce a call to a Chrome API function which will mean it won't be
possible to use the JS as a bookmarklet anyway.