Add integration test using Capybara and Poltergeist
We've taken a copy of the FreeAgent unexplained transaction HTML, we're
serving it using Webrick and testing our JavaScript by executing it
within that page.

*NOTE.* The unexplained transaction HTML still contains references to
external JavaScript and CSS but we plan to do something with those
separately.

The JavaScript needs the URL of the rules which we're supplying by
setting `window.rulesUrl`. If this isn't set (i.e. when the JavaScript
is executed within the context of the Chrome extension) then it'll
retrieve the URL from Chrome storage as before.

We've created a single sample rule in `spec/fixtures/test-rules.json`
that we're expecting to match our unexplained transaction.

We've introduced a `test:integration` Rake task and added that to the
default Rake task.

The integration test starts a WEBrick server in a thread so that we can
serve our fixture files. The server is shutdown in the test teardown.
We're assigning the `Thread.new` instance to an instance variable to
avoid the case where it might be garbage collected too early.