Display message when no attachment is required
I think it's better to explicit state whether an attachment is required
or not. We already had a message to explain that an attachment was
required, and we now have one to explain when it's unnecessary.

* Load jQuery and jQuery UI in jasmine.yml configuration as our
JavaScript relies on both of them.

* Stub `$.getJSON` so that we can programmatically inject the rules
we're interested in.

* Supply the `fail` function as a no-op to avoid an 'undefined
function' error.

* Use jQuery to create the HTML fixture required by the unit test. I'm
not convinced this is the best way but it's good enough.

Fixes #1