Provide a keyboard shortcut to trigger the extension page action.
I've set the default to be `Ctrl+E` which is equivalent to `Command+E` on a
Mac, but note that you can change it by clicking the "Keyboard Shortcuts" link
at the bottom of the `chrome://extensions` page.

Note that this keyboard shortcut is only active when the extension page action
icon is visible.