Only attempt to explain transaction once.
Previously we were allowing multiple attempts to explain the transaction, but
this led to problems where we were appending text, because we'd keep appending
the text over and over again.

I could've tried to fix this by trying to determine whether we'd already
appended the text or not in order to decide whether to append the text.
However, even this isn't quite right, because theoretically the rules could've
changed in between the first attempt and the second attempt and so the message
might be different or non-existent the second time around.

Although it's not ideal, it feels as if preventing multiple attempts is the
best solution for now.