Change version of Chrome extension from 1.0 to 0.1.
The Chrome extension doesn't feel "ready" enough to merit giving it a version
of 1.0.

Although it might be confusing to anyone who already has the Chrome extension
installed, since we haven't yet published or publicized the extension, it's
probably safe to do this and definitely better to do it now than later.