Fix build.
The build started failing on Travis CI [1]. The problem was that neither
the jQuery nor the jQuery UI libraries were being loaded into the page.
I'm not sure why this has only just become a problem.

In normal operation of the Chrome extension these libraries are loaded via
the extension's manifest file, but in the integration specs we're taking a
bit of a shortcut, so we need to load them manually.

[1]: https://travis-ci.org/freerange/freeagent_transaction_explainer/builds/39270289