Fail fast if element missing or ambiguous.
We want to make the Chrome Extension more robust against changes to the
FreeAgent bank transaction page HTML. These changes mean that an error is
thrown if an element is missing or ambiguous. Errors are caught and logged
and an alert is displayed.

I decided to display an alert so as not to rely on the presence of any element
in the page. I'm afraid I haven't added any specific tests, but I did have to
extend the fixture for the jasmine spec to avoid the new errors being thrown.

Fixes #39.