Fix link to properly upgraded extension in CWS.
First time around, I was confused by the Chrome Web Store Developer Dashboard
user interface and so I created a whole new "extension". I've now managed to
publish an upgrade to the original extension which has the advantage that
its easier for people to upgrade the extension in their Chrome browser. In
fact I think it might even happen automatically.