Update travis.yml to use precise distribution
The tests are currently failing on the trusty distribution. I believe
this is because of the upgrade to PhantomJS 2 as I'm also seeing the
tests fail locally with this version of PhantomJS.

Ideally we'd get the tests passing using the later version of PhantomJS
but I'm hoping this change will at least ensure they're passing on
Travis again.