Fix JS errors in integration spec
I was seeing a couple of JS errors like this one:

    TypeError: undefined is not an object (evaluating 'n.info.beacon')
    TypeError: undefined is not an object (evaluating 'n.info.beacon')
        at http://127.0.0.1:59835/freeagent-unexplained-transaction_files/nr-515.min.js:1 in t
        at http://127.0.0.1:59835/freeagent-unexplained-transaction_files/nr-515.min.js:1
        at http://127.0.0.1:59835/freeagent-unexplained-transaction_files/nr-515.min.js:1
        at http://127.0.0.1:59835/freeagent-unexplained-transaction_files/nr-515.min.js:1 in r
        at http://127.0.0.1:59835/freeagent-unexplained-transaction_files/nr-515.min.js:1
    # ./spec/integration/integration_spec.rb:23:in `block (2 levels) in
    # <top (required)>'

I assume the newer version of Poltergeist is somehow being stricter
about catching JS errors and that's why this has started happening.

The errors were occurring in a JS file which isn't obviously relevant to
the core functionality on the page. Although it's not ideal, I decided
that the most pragmatic solution was to avoid loading the file and to
remove it from the fixture directory.

In any case, we'd probably be better off obtaining a more up-to-date
version of the FreeAgent unexplained transaction page, rather than
spending time fixing this problem.