Handle rules with VAT set to "Amount...", i.e. manual
This is really ugly. I couldn't find a way to select the VAT option as
if a human was doing it, i.e. to avoid having to trigger the
side-effects manually. However, there's some magic stuff going on in the
FreeAgent JS, possibly relating to data attributes and/or a shadow DOM.
Anyway this is the best I could do for now and I think it's better than
nothing.