Fix highlighting of elements changed by extension
The jQuery UI highlighting hasn't worked for quite some time. I think
this is because of the CSS styles of the relevant elements and their
parents. I've fixed this by just setting the border-color to red. I
don't think it matters that this is not then un-set. In fact I think
it's possibly better that it isn't un-set.